+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/quizread/QuizSearcher.kt
@@ -1,6 +1,5 @@
 package com.grepp.quizy.quiz.domain.quizread
 
-import com.grepp.quizy.quiz.domain.user.UserId
 import org.springframework.stereotype.Component
 
 @Component
