+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/quizread/UserQuizSearchService.kt
@@ -1,7 +1,7 @@
 package com.grepp.quizy.quiz.domain.quizread
 
 import com.grepp.quizy.quiz.domain.global.dto.Slice
-import com.grepp.quizy.quiz.domain.user.UserId
+import com.grepp.quizy.quiz.domain.useranswer.UserId
 import org.springframework.stereotype.Service
 
 @Service
