+++ quiz-service/quiz-infra/src/main/kotlin/com/grepp/quizy/quiz/infra/quizread/repository/QuizSearchRepositoryAdapter.kt
@@ -2,12 +2,10 @@ package com.grepp.quizy.quiz.infra.quizread.repository
 
 import com.grepp.quizy.quiz.domain.global.dto.Slice
 import com.grepp.quizy.quiz.domain.quizread.*
-import com.grepp.quizy.quiz.domain.user.UserId
 import com.grepp.quizy.quiz.infra.quizread.document.QuizDomainFactory
 import com.grepp.quizy.quiz.infra.quizread.document.SortField
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Sort
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Repository
 
 @Repository
